feat: Add title, icon & remove "create-react-app" stuff
As any true application it should have a icon and a name, not "React App" as "create-react-app" generates. So I overwrite favicon.ico with a Iconfinder books icon, which is pretty and cover other dimensions, rename index.html title and applied this on manifest.json.

I've also remove logo, default App's style, test and content.