feat: Add React Context & ShelvesOptions component
I'm using a lot of components. It's hard to deep track up an event and
track down states.

Because of the buzz about new React's Context API and how it can be useful to
made this state changes I decided to test and liked the result. I've made a
StoreController that wraps components with Provided state and consume it using
StoreConsumer component.

I've also create a ShelvesOptions component that maps shelves into ShelfOption
component. ShelfOption receives book and a shelf, it define if option is the
active one and handles onClick.